Add test for trees being updated
We had a bug where when a re-org happened, the note / nullifier trees
were trimmed too far back. The forking points notes / nullifiers were
also removed. This adds a test to make sure that the trees are
updated on a linear add, but also properly pruned and re-added during a
re-org.

This replaces the test that was in the file previously. Fixtures have
been updated.