Add backup command
This new command will tar your data folder and upload to a specified S3
bucket. It's a hidden command because we're using it to test during test
net, and isn't really intended to be used by end users. It requires
'tar' and 'curl' programs to be installed on your computer.