Remove NetworkBridge
The idea behind the network bridge, is that we'd move Captain entirely
inside of a worker, and then serialize messages across. We've changed
strategies and are moving toward a WorkerPool approach, where we spin up
workers based on the number of cores, and move tasks that will freeze up
the main thread onto those workers.