Group primitive classes with their equivilents.
So the purpose of this commit, is that in the past we had Ironfish
implementations of things in Strategy. That means, that block.ts didn't
actually contain the block, but strategy.ts contained the block. For
each concept (block), I moved the Ironfish version into the
corresponding file. So block.ts contains

 * BlockSerde (serializer)
 * IronfishBlock
 * Block class

Soon when we delete Strategy, only the block class will remain. We had
the strategy so we could have other implementation of our coin. The
generics were a bad design idea and were slowly unravelling them to
remove them by materializing our concepts.