Fix ancestor search considering islands cache hit (#114)
This fixes a bug where you get stuck trying to sync an orphan block
island. It's because some how orphan block islands are being added to
the database, even though were syncing left to right now. The issue is
once this happens, you will find an ancestor on the island, and try to
sync from there. This is slower, but will now not consider island blocks
a cache hit.

Note that this can all be deleted once we remove graphs, remove adding
islands from blockchain.