Refactor block syncing (#100)
* Add New BlockSyncer

This replaces the old block syncer. The new strategy syncs left to right
using the ancestor block of the remote peer. It figures out who to sync
from by looking at who has the most work.

* Fix peer.head error

we were removing the head when we exhaust a peers list of blocks to
download, this should fix that.

* Enforce consensus max request limit

* Add first tests for Syncer

 - stops syncing when peer disconnects
 - should stop syncing on error
 - should start and stop syncer
 - should sync from peer, and stop when finished

* Fix ancestor search considering islands cache hit (#114)

This fixes a bug where you get stuck trying to sync an orphan block
island. It's because some how orphan block islands are being added to
the database, even though were syncing left to right now. The issue is
once this happens, you will find an ancestor on the island, and try to
sync from there. This is slower, but will now not consider island blocks
a cache hit.

Note that this can all be deleted once we remove graphs, remove adding
islands from blockchain.

* Dont sync orphan while syncing another chain

* Handle errors in synching chain and surface them

* Document why we return stack in renderError

* Clarify consensus message

* Delete unused blocks message

* Removed unused enum value

* Dont add orphan blocks during syncing (#117)

This is a temporary measure because there is a bug in the graphs of the
chain where the graphs become disconnected. We shouldnt add orphans
anyway, since we are gutting that logic. It'll be deleted soon, but this
makes the chain graphs more stable by only adding non islands.

* Clarify TARGET_BLOCK_TIME_MS

* Abort gossip by having handlers return boolean (#119)

Before there was no way to abort the gossip, now we can abort the gossip
by returning true or false from a gossip handler.