Various changes to syncer during debugger (#172)
* Add more detail to invalid message errors

* Add peer name whitelist for syncing

Now you can add peer names to the syncer whitelist for debugging
problems and only sync from those peers. All gossips are dropped as well
from other peers.

* Sync every 2 back from our head instead of 1 back

I made this change because at best we only go 45 seconds back on a
linear search, but if we go back 2 we go back 90 seconds. This gives us
more of a chance to get a hit on the linear search.

* Measure speed of duplicates and orphan adds

The syncer speed only included added blocks, but now includes all blocks
even if they are orphans or duplicates. You can still see the add speed
of only added blocks using the blockchain add meter. Both are displayed
in the status CLI.