Add prefixes for repair block command (#177)
* Fix meter not reset current counters on stop

When the meter stops, it needs to reset these counters. If you start
again, it'll consider what ever was left in these buffers to be the last
sample which are skewed by a massive time sample now form when it was
last running.

* Add a way to reset meter

This resets all the average and makes them start back from 0

* Fixed progress bar interface to match library

* Clean up checkTreeMatchesHeaviest

The logs were pretty sloppy here.