Commit addBlock state after each DB operation
We had one transaction / lock around blockchain.addBlock(), that if
rolled back, would leave dirty in memory state. To solve that
temporarily, we added a new IDatabaseTransaction.update() function which
updates the database but does not release the lock. This allows us to
commit state to the DB without releasing locks.

Alternatively, we can look at other solutions in the future where
transactions can be moved downwards to the things that need to be
atomic.