Dont gossip blocks when we don't add the block (#190)
We should only gossip blocks when we added them, because if we gossip a
block, it signals to others that we also have that block on our chain.
It's useful because if someone gets our gossip block, and its an orphan,
they will want to download the chain from us to that oprhan. So if we
gossip without adding, then that won't work.

Many coins distinguish this as announce vs broadcast, where announce
tells people we have it, and broadcast gossips it without indicating we
have it or not. Announce sends entire hash, broadcast sends entire
block.

If you look at the code, there is one situation here where we have the
block but we don't gossip it, and that's when it's a duplicate and we
already have it. At that point we don't regossip it out. We can
determine if we should be gossiping it out later in this case.