Handle gossip blocks when we dont have a syncer (#208)
Before we didn't handle gossip blocks when we were syncing, now we can
do it if we don't have a syncer. This can help fix when we're stuck ina
small network without a lot of blocks being produced. If someone manages
to get a block in the network, and were not synced, this will now kick
off the syncing process