Make iterate errors optional
So these errors are used to catch chain corruption, but now that we
fixed chain corruption bugs they are triggering in valid instances. Our
account head updates, but doesn't lock so the chain can be modified
while were iterating due to a reorg. This will cause accounts to
crash because these validation check that we should be able to validate
the entire expected range.

So in accounts it's fine if we have to abort due to a reorg. Honestly
long term we can consider removing these errors once this is polished
and we can be sure there are no more chain corruption issues for sure.