Fix ironfish reset by not checking upgrade
There are instances where we want to open the DB without running any
upgrades. This introduces a new upgrade argument to accounts and
blockchain open() methods. It only checks for mandatory upgrade when
true is passed.

Later we'll remove this in favor of a better system when we have time
and have added DB checks.