Remove transaction scope and type
This just removes the two arguments to IDatabase.transaction()
and IDatabase.withTransaction called scope and type. You may
see some other changes because it reduces 1 level of indent,
so now the code. lines all shift.

These were never used. They were going to be used for the web when we
introduce IndexDB, but it's not necessary because we can just use the
code the way it is now, there.