Add database version and migration errors
This adds a new mechanism so we can throw errors when database schemas
change. It will now enforced the DB version and a reset in a less hacky
way. A simple global database version is stored in the DB, and specified
to DB.upgrade(version).

If the version mismatches, then an error is thrown. If you are writing a
migration, you can avoid calling this method to run your migration logic
on old code.