Delete TestStrategy and usages
I removed TestStrategy and all usages. We converted over tests to test
our actual code so we don't need it anymore. Overall it was a bad
idea to begin with because our actual code wasn't being tested, and once
we test that it no longer serves a purpose. On top of that, once we
tested our code we found crashes that the TestStrategy tests didn't
catch.