Remove generics
Generics were originally designed to allow our coin to be composable,
but added a lot of extra complexity to the code base. We've been
working for the past few months to remove some complexity, including
materializing generics with the exception of non continer types like
MerkleTree. As well as remove TestStrategy which used the generics.

The ultimate goal is to make the code base easier to understand to
developers and the open source community.