Fix account rescan
Rescan was broken for a few reasons.  It didn't attach to an existing
rescan properly, it also didn't indicate well that it was retaching when
I fixed that. The rescan code itself was iterating the wrong way which
resulted in a crash when it couldn't find the next block. The fixes
are..

 * Added startedAt to rescan so existing attachments can see how long
 the existing rescan has been running
 * Fixed iteration direction, it was going left from the genesis block
 instead of right causing it to immediately abort
 * Attaching to rescan was broken because it didn't await the actual RPC
 command causing it to exit instanstly