Add API block syncer
This adds a new RPC command that follows the chain starting at an
arbitrary head, if any. Once it connects, and blocks are added, the
follow RPC command will stream back the blocks and the block event based
on the chain operations that occur. It only streams back blocks on the
main chain. This will be changed later.

The new command now will accept and API endpoint and token for an Ironfish
API and POST those blocks to the endpoint. That API can be found here,
https://github.com/iron-fish/ironfish-api