Manually enable JobAbortedError per job
This is a new feature that is causing large errors to be printed when
jobs don't handle the JobAbortedException. In the past when you were
shutting down these jobs would just hang and cause problems. Aborting is
a good behavioural change, but we haven't yet gone through and
upgraded all job call sites.

This adds a new bool to each job that you can turn on depending on the
job if you want to handle the job aborted error.