Fix balance.test.ts (#380)
This test was broken for two reasons. First, sdk.connectRpc() wasn't
being mocked, which was simple enough.

The second issue is that the test
was testing the formatting matched the output of the command, but the
spacing was inconsistent. Now, it only tests that the balance numbers
get output and doesn't care about how the message is rendered.