Removed dependency on NodeJS types
This was causing errors like this:

> src/metrics/meter.ts:71:5 - error TS2322: Type 'number' is not assignable to type 'Timeout | null'.
>      this._interval = setInterval(() => this.update(), this._intervalMs)

The reason is because the ironfish project is cross platform compatible,
and cannot be dependent on compiling with node types. We accidently used
the node type Timeout, which returns a different type than number
in the browser. Because of that, to make it cross compatible we need to
reference the generic instead so we don't hard code it to reference one
single platforms type.