Build ironfish-rust as native NodeJS addon
Adds an ironfish-rust-nodejs package to the repo that uses Neon to build
ironfish-rust as a native NodeJS addon instead of a WASM bundle. This
should significantly improve performance on all machines.

Also renames ironfish-wasm to ironfish-rust-wasm, and cleans up a lot of
code that was referring to wasm specifically in the ironfish package.