Remove isWorker/broadcastWorker code
The isWorker/broadcastWorker code isn't used anymore now that our
API node architecture has been changed, so we can remove the code
to simplify our networking a bit.

This is a breaking change for anyone starting their node with the
worker flag, but we're okay with breaking compatibility during the
testnet phase.