Don't rebroadcast invalid transactions (#409)
* Don't rebroadcast invalid transactions

This prevents a very specific problem where we create a transaction on a
fork that is no longer valid on the main chain. The issue is that we'll
permanently rebroadcast this transaction but no one will think it's
valid. This causes the transaction DDOS mechanism that is taking down
the network.

This is NOT the permanent solution because it will check the
transaction forever, and never unmark the notes as spent that were spent
on this fork.

The correct solution is...

1. Transaction expiration to eventually remove this transaction
2. Properly unmark notes as spent by deleting the transaction that was
   created on a fork when the block reorg occurs.

These fixes are coming in later patches.

* Fixed test name and remove only

* Fix orthagonal comment typo

* Remove out-dated comment

* Fix lint