Change max synced age ms to use block time
This was originally meant to be 4x block time, but since we changed the
block time, this is out of sync. Were also making this 12 hours
temporarily because of all the long blocks to mine issue. We'll adjust
this and reduce it again if we find that too maqny forks exist.