Allow syncer to sync with peers even when synced
This issue was introduced in
https://github.com/iron-fish/ironfish/pull/647 which is moved our sync
time definition back 10 hours. This solved another issue, but now when
people become synced they are suddenly stuck because the syncer won't
run anymore.

This is an experimental change, but it allows the syncer to run at any
time when someone has more work than us. Remember, that gossip also
still kicks off a sync as well.