Fix banning peers for expired transactions
There is a bug in that we verify a transaction expiration based on the
current chain head, rather than on the block the transaction was added
to. This means if you sync an old fork from someone you are nearly
guaranteed to ban them.

Consider this chain

G -> A1 -> A2
  -> B1

If you have a transaction B1 that expires at height 2 (expiration is
inclusive) then sync that chain from another node, you'll ban them. From
your perspective, B1 has an expired transaction because you have block
A2 as your head.

The fix is to validate transactions based on their block, and not your
current head.