Remove account and config RPC calls from TCP
This removes RPC routes in the accounts and config namespace from being
mounted in the TCP adapter. This is to prevent a an attack where
users can have their graffiti changed. This is a short term fix because
the real fix is to add RPC token authentication.

Accessing one of these routes now with TCP will display one of the
following errors.