Dont set config overrides unless not default value (#910)
We found a bug where the base command was setting config overrides for
values that are the same as the flags default value. This causes the
config file to be ignored.

Now, we only set the config override if it's different from the default,
which allows the config not to be over-ridden with the flag default.