Allow version to be driven by lib consumer
This makes a change so that previously, the version sent out everywhere
and the agent string was driven by the library. This doesn't tell the
full story, because the real version is the person who is consuming the
library and potentially writing custom code on top of the
library.

This lets users who consume the library override the package and produce
their own agent and version library using the new Package type that you
can produce from a Package.json file and pass down into the node.