Chain should reject invalid with the same reason
We have a mechanism in our blockchain such that when we validate an
invalid block, we remember that block hash and reject that block if we
ever see another block with that hash again as an optimization.

The issue is that future rejects of that same block reject with reason
"error" which is very confusing. It should remember the rejection reason
and reject with the original reason.