Fix rescan when account head is on a fork (#937)
* Fix rescan when account head is on a fork

This fixes the issue outlined in github issue #880. This is happening because of
how we handle rescans.

1. When a rescan occurs, the wallet will never update itself until the rescan
   completes.
2. When the node starts and a rescan has started it will always try to rescan
   first

This means that because the rescan is broken, the wallet is stuck in a broken
state permanently. Let's now dive into why the rescan is broken.

```
G -> A1 -> A2
        -> B2 -> B3

Accounts Head: A2
Chain Head:    B3
```

When we do a rescan, we try to iterate from the genesis block (G) to the
accounts head (A2), but the block chain has re-orged to B3, the newer heaviest
chain. If the accounts head were to update, it would reorganize as well to B3.
Unfortunately, once a scan has started the accounts will wait for the rescan to
finish before updating the accounts head from A2 to B3.

Now the way we try to iterate from G -> A2 is that we use the heaviest chain
iteration. That means we're going to iterate in this order: `G -> A1 -> B2`
before realizing were not on target and fail with the error that's in the
ticket. It's now unrecoverable because there's no way out of this since accounts
is stuck at A2.

The fixes I chose were inspired by Bitcoin, which chose not to solve the problem
exactly. The worst case outcome are that if you are importing an account with a
transction on a fork, you'll potentially think that transaction is added to a
block on a fork, even if it never gets added to a block on the main chain.

The fixes are two fold:
1. Don't error when the account head is unreachable
2. Don't let iterateTo jump chains if we use iteratetoHashes() during a

* Added a test

* Fix tests

* Fix iteration count