Move hostsStore instantiation to node; fix tests
Having the hostsStore instantiation in the AddressManager
requires an await on the load() for the backing file store.
This leads to a floating promise, which should be addressed.

As one cannot use async/await inside a class constructor, the
only other choices are to do something similar to what we already do
in ironfish/src/node.ts in which we make a static async init()
function, or to instantiate and load the hosts store earlier in the
node setup and pass it to PeerManager for use in creating AddressManager.

This PR is already fairly large, so I am not sure that adding init
functions at every necessary point is the way to go when we can instead
load the hosts store at the same time as the other file stores. In the
future, we can create the aforementioned init() functions if we
deem it necessary.