Add BlockTemplateStream RPC endpoint
The new BlockTemplateStream RPC endpoint will provide all of the pieces necessary for miners and pools to construct and modify block headers however they need. This is the first step to enabling pools by enabling ways for miners to modify, for example, the block header's graffiti to ensure work is not being duplicated between miners.