Documentation updating (#1052)
* Doc update 1

Remove test:slow
Add referencing bugs fixed in the PR description.

* Update graph explorer readme.md

* Update locations of where test: commands can be run

* Add request to add bugs being fixed to PRs

* Update graph explorer readme

* Update CONTRIBUTING.md

* Bump ironfish-cli to 0.1.22

* Update to latest napi-rs to fix BigInt usage (#1014)

* Update RpcClient routes to use ApiNamespace enum (#1013)

Minor change to RpcClient to utilize the ApiNamespace enum for route definition. Since we have a non-stringly-typed enum already, using it here just minimizes the risk of a typo that breaks a route.

* Move telemetry tags to fields (#1026)

We are doing this to reduce cardinality. Indexes are created from the
combination of measurement and tags. This means every node session ends
up getting it's own indexes which blows out our cardinality and causes
causes InfluxDB to stop processing data. You can read more about this
issue here:

https://docs.influxdata.com/influxdb/cloud/write-data/best-practices/resolve-high-cardinalityg

* Fix ordering issue in creating telemetry node id (#1027)

The issue is that Node.Init() happens before the telemetry node id is
created for the first time. We simply move that up, and configure it on
the SDK itself which gets passed to the constructed node.

* fix miners:mined scanning status (#994)

* Return a Buffer instead of a NativeNote from decryption functions (#1025)

* feat(ironfish): Move metrics monitor inside telemetry (#1029)

* feat(ironfish): Submit peer metrics in telemetry  (#1030)

* feat(ironfish): Move metrics monitor inside telemetry

* feat(ironfish): Submit peer metrics in telemetry

* Bump version to 0.1.23 (#1028)

* Add telemetry status to status command (#1036)

* feat(ironfish): Update telemetry flush and metrics intervals to 5m (#1038)

* feat(ironfish): Update telemetry flush interval from 5s to 60s

* Update to 5m

* Update METRICS_INTERVAL

* fix(ironfish): Pass in metrics and remove peer network from telemetry (#1033)

* fix(ironfish): Pass in metrics and remove peer network from telemetry

* s/peersCount/p2p_PeersCount

* Change getStatus() RPC to use new metrics gauge (#1045)

* v0.1.24 (#1044)

* Doc update 1

Remove test:slow
Add referencing bugs fixed in the PR description.

* Update graph explorer readme.md

* Update locations of where test: commands can be run

* Add request to add bugs being fixed to PRs

* Update graph explorer readme

* Update CONTRIBUTING.md