Make IpcAdapter.stop() destroy all sockets
This bug is surfacing because rpc.Stop() is hanging forever, because
underneath is stops all the current adapters. When the IpcAdapter tries
to stop, it waits for all the connections to end. The problem is that
IpcServer.stop() doesn't termiante the connections. This new change will
terminate the underlying connections causing all the connections to end,
thus stopping IpcAdapter.stop() from hanging.

This was first documented in this PR, which a hack was added just to
ignore RPC stop in this case.

https://github.com/iron-fish/ironfish/pull/1011