Move TCP-connection logs to debug, add a miner log
Logging TCP connections is noisy due to port scans, etc. Moving them to debug allows them to be silenced for the primary use case, while still giving access to them when needed. Adding an info level log for when a miner has successfully subscribed, since that's primarily what a pool will be interested in seeing.