Update to proper sqlite3 version (#1234)
sqlite3 package is under new management, which broke the downloads of the pre-compiled binaries. The upside is that they finally released an updated version with the changes we were using, so now we can use the regular npm version. See https://github.com/TryGhost/node-sqlite3/issues/1572 for more information about the change in ownership.