IRO-1940: define IronfishClient, move IronfishRpcClient down one level of abstraction
implementing a TCP client using the native 'net' module will introduce shared
logic with the existing IPC client. the `IronfishMemoryClient` does not have the
same network connection logic that TCP and ICP do, so an additional level of
abstraction is useful for sharing logic between TCP and ICP implementations.

Changes:
- add client.ts, define `IronfishClient`
- move implementation of `IronfishRpcClient` into `IronfishClient`, minus
  abstract functions for connect/close
- change references from `IronfishRpcClient` to `IronfishClient`