Fix incorrectly kicking peers if we have the TX
There is a bug where users are being kicked from connections if they
rebroadcast transactions to us that we already have. The issue is that
wallest are programmed to rebroadcast transactions if we don't accept
them up until their expiry time. However, acceptTransaction() returns
false if we already have the transaction. We also punish any user
for if acceptTransaction() returns false. You can see the issue
here.