feat: implements secure adapter and client for TLS over TCP
extends TcpAdapter and IronfishTcpClient to establish TLS connection. expects
the node's private key and a self-signed cert to be located in the node's data
dir. the client is configured to allow the unauthorized, self-signed cert from
the node.

includes changes:
- refactors IronfishTcpClient and tests to make it possible to extend the class
  for TLS connections -- tls.Socket is a subclass of net.Socket, but doesn't
        implement its own 'connect' method, so the connection flow doesn't quite
        match the prior implementation
- adds config options for tlsKey, tlsCert, and enableRpcTls

it might be helpful/necessary to provide a script and documentation of how to
generate a key and cert for the node to use TLS. we'll also need to make some
changes to the Dockerfile to get a key and cert in the right place in the
containerized application.