Delete disconnecting peers for non acceptable tx (#1517)
The main thing at this point, is that this would disconnect people who
were creating transactions on a fork. I don't think we'll need to bring
this back, it also seems like this was resolved by ther other changes
mat adde, to not accept mempool entires for conflicting or invalid
tranaactions. We should instead build confirmation and fix forking.