feat: adds script for generating key and cert for tls
starting a tcp server over TLS requires a private key and a public cert for the
server. this script executes the steps required to generate a private key and a
self-signed certificate for the server.

- uses default private key path of 'node-key.pem'
- uses default certificate path of 'node-cert.pem'
- uses default cert subject of '/O=Iron Fish', where 'O' refers to
  'organization'

moving the key and cert into the node's data directory isn't within the purview
of this script.