feat: adds a message buffer to the stratum client
protects against reading partial messages from the TCP socket stream.

stores any partial messages read from the socket in the message buffer and
concatenates the next read onto the message buffer.

background:
tcp sockets are modeled on streams of data, and the data that the server writes
to the socket may be broken apart during transmission. the reader is guaranteed
to receive data in the same order that the writer sends it, but not necessarily
in a contiguous chunk.

- uses last index of message delimiter ('\n') to identify beginning of any
  partial messages
- clears message buffer on socket disconnect