feat: implements secure adapter and client for TLS over TCP (#1508)
* feat: implements secure adapter and client for TLS over TCP

extends TcpAdapter and IronfishTcpClient to establish TLS connection. expects
the node's private key and a self-signed cert to be located in the node's data
dir. the client is configured to allow the unauthorized, self-signed cert from
the node.

includes changes:
- refactors IronfishTcpClient and tests to make it possible to extend the class
  for TLS connections -- tls.Socket is a subclass of net.Socket, but doesn't
        implement its own 'connect' method, so the connection flow doesn't quite
        match the prior implementation
- adds config options for tlsKey, tlsCert, and enableRpcTls

it might be helpful/necessary to provide a script and documentation of how to
generate a key and cert for the node to use TLS. we'll also need to make some
changes to the Dockerfile to get a key and cert in the right place in the
containerized application.

* requires the node cert to authorize TLS connection

using the 'rejectUnauthorized: false' does encrypt traffic, but allows any cert.
requiring the node's cert from the client side ensures that the server's cert
matches what the client expects.

* fix: removes unnecessary 'rejectUnauthorized' config from server

* read key and cert in sdk, pass to adapter and client

* fixes linting error

* changes TLS client to not perform certificate validation

* removes unused logger imports