feat: adds a config variable for pool payout expiration, sets default (#1587)
we believe that some payout transactions from the pool are expiring. we have a
default expiry of 20 blocks for such transactions. these changes increase the
expiry to 60. with one minute block times this should be roughly equivalent to
half of the duration of the payout interval. adding a config value for this
field will help to change it without a code change if necessary.