fix: restore enableNativeRpcTcpAdapter option, add option for client
we've recently made several changes in the rpc layer:
- native TCP adapter
- native TCP client
- TLS adapter/client

it may be useful to be able to test these features separately. this commit
restores an option for enabling the native TCP adapter and introduces an option
for enabling the native TCP client so that we may enable and test the adapter
and client separately.