Increase the max queue default size for the worker pool (#1608)
The original value of 200 is not ideal anymore, as we have code that can kick off 300 jobs at a time, which would cause other jobs to be skipped. One example of this is when a block comes in, we kick off a `VerifyTransaction` job for every transaction in the block, which currently can be 301. If a gossiped transaction comes in at the same time, we are likely to not gossip that transaction when checking if the queue is saturated, even though it will no longer be momentarily. With a value of 500, we can expect a block to come in, with some leeway for other jobs before we count it as saturated.