feat: adds message buffering to rpc tcp sockets
a tcp socket is modeled as a stream of bytes. we use tcp sockets to send and
receive json-formatted messages, but there is no guarantee that a tcp segment
will contain a full message or messages. it's possible, especially with larger
messages, that we read data from the socket in the middle of a message.
attempting to parse such partial messages results in json parsing errors.

- implements a MessageBuffer class to encapsulate the logic required
for buffering partial messages read from a socket

- adds a MessageBuffer to TcpClient
- adds a MessageBuffer to each SocketClient
- updates 'data' event listeners to use MessageBuffers
- adds MESSAGE_DELIMITER to protocol definitions