PeerNetwork should sync transactions regardless of mempool accepting it (#1624)
* PeerNetwork should sync transactions regardless of mempool accepting it

This fixes issues where accounts can become out of sync when used on multiple machines, primarily due to creating transactions on all of them.

* Move verification from mempool to onNewTransaction

This changes the `peerNetwork` function `onNewTransaction` to explicitly verify transactions, instead of implicitly relying on the mempool to do it while accepting transactions. Since we're doing it here, we can skip doing it during the mempool `acceptTransaction` function.

Also changes the `verifier` function `verifyTransaction` to `verifyTransactionContextual`, and splitting out a new function from that called `verifyTransactionNoncontextual`. While this is accruing a bit of tech debt by creating more functions for verifying transactions in the verifier, it is a debt we are going to be paying down soon when refactoring `verifier`.