decouples StratumClient from MiningPoolMiner
background:
the StraumClient and MiningPoolMiner are tightly coupled -- an instance of one
type holds a reference to an instance of the other type. this coupling makes it
difficult to use StratumClient without a MiningPoolMiner. a developer might want
to do this if they were to implement their own miner, or if they wanted to use
the StratumClient to receive messages outside of a mining context (e.g., to get
status information from the pool).

- removes miner property from StratumClient
- modifies StratumClient to emit message bodies after validating