decouples StratumClient from MiningPoolMiner (#1648)
* decouples StratumClient from MiningPoolMiner

background:
the StraumClient and MiningPoolMiner are tightly coupled -- an instance of one
type holds a reference to an instance of the other type. this coupling makes it
difficult to use StratumClient without a MiningPoolMiner. a developer might want
to do this if they were to implement their own miner, or if they wanted to use
the StratumClient to receive messages outside of a mining context (e.g., to get
status information from the pool).

- removes miner property from StratumClient
- modifies StratumClient to emit message bodies after validating

* removes check of stratum connection

- miner now awaits stratum connection, so stratum client will be connected if
  miner has started

* moves message event fields to top of StratumClient class

* fixes handling of stratum connection

- removes blocking in StratumClient.start and MiningPoolMiner.start
- adds onConnected event to StratumClient to notify listeners
- restores check of stratum connection in mining loop

* removes await or miner start in miner start cli command