Abort rescan if aborted after waiting for head
When were rescanning, we start a scan, then wait for the lock on the
updateHead() method. However, this may take a long time. It's possible
the head updating was aborted because the node is shutting down. The
issue is that we don't also check if the scan was aborted as soon as
the updateHead() lock is released.

This means we could start scanning when the node is shutting down which
will prevent it from shutting down properly.