Add Database Migration System (#1793)
* Add migration system and blockchain migration

This adds a new database migration that forces users to run (potentially) long
running migration operations on their node before the node will start. This does
have an impact on user experience so I'll document that below.

I've repurposed the database version to now be used by the migration system, and
no longer used by the database layer itself. Migrations are only meant to
migrate one single database, because we can't guarantee atomic transactions over
multiple leveldb instances anyway. Because of that, you should never write a
migration that requires atomic writes between multiple databases anyway.

* Get tests passing

* Fixed lint errors

* Fixed slow test

* Fixed command descriptions

* Pull migration name from file path

* Fix debug running when DB is not upgraded.

* Remove default export

* Fix parallel applied checking

You cannot check if a migration is applied in parllel because if you
buse the same DB they'll get a locked error and it will crash the
migrations since they are both opening the same DB and checking it.