Disallow importing accounts with the same keys
There isn't a valid reason to have the same account imported multiple times, and it is a drain of overall node performance since each note needs to be decrypted twice for no good reason. This also allows further optimizations with decrypting notes such that we could stop trying to decrypt a note for accounts once we've found one match, which can be a sizable performance increase for nodes with multiple accounts.