Fix a bug where tests were timing out
The default timeout value of jest is set to only 5 seconds. This
period seems unsufficient for some tests that appearantly take
longer to execute. This commit changes the timeout to 32 seconds,
which should suffice for the current tests. An improvment would be
to make this value get set dynamically according to the specs of the
computer running the tests.