Fixed a bug where not all errors were handled properly
The issue #1792 reported a bug where `ironfish debug' command
had an exception message shown. After investigating the code, it
turned out that not all exceptions were handled by the catch block,
which left the `dbOpen' flag variable set to to (when the database
may not be actually open). This commit also contains a refactoring of
the output collection functions.