adds 'temp' directory in dataDir, uses it in snapshot commands (#1912)
* removes database name from snapshot on unzip

when using `service:snapshot` to create a snapshot, the resulting zip file
contains a directory matching the database name. if a user later imports that
snapshot then the database name in the snapshot shouldn't need to match the name
that their node uses.

- uses `strip` to remove the database directory on unzip
- creates a 'snapshot' directory and unzips the contents of the snapshot there
  before moving to the user's database directory

- fixes snapshot to create directorys on 'path', importSnapshot to resolve
  'path'

* adds 'temp' directory in dataDir, uses it in snapshot commands

snapshots are written to and downloaded to a random directory within the
system's `$TEMPDIR`. this can create duplicate snapshots and downloads and makes
the snapshots difficult to find.

- adds `tempDir` getter to config
- replaces os tempDir with node tempDir in snapshot commands